+++ camayoc/utils.py
@@ -6,6 +6,8 @@ import os
 import shutil
 import tempfile
 import uuid
+import json
+import base64
 from urllib.parse import urlunparse
 
 from camayoc import exceptions
@@ -77,3 +79,21 @@ def isolated_filesystem():
             shutil.rmtree(path)
         except (OSError, IOError):
             pass
+
+
+def create_identity(account_number, org_id=None):
+    """Base64-encoded JSON identity."""
+    identity = {'identity': {'account_number': account_number}}
+    if org_id is not None:
+        identity['identity']['internal'] = {'org_id': org_id}
+    identity = json.dumps(identity)
+    identity = base64.standard_b64encode(identity.encode('ascii'))
+    identity = identity.decode('ascii')
+    return identity
+
+
+def create_x_rh_identity(account_number, org_id=None):
+    """"Base64-encoded JSON identity header provided by 3Scale."""
+    identity = create_identity(account_number, org_id)
+    x_rh_identity = {'x-rh-identity': identity}
+    return x_rh_identity
